Public API

NeperStructureGenerator.mesh — Function

mesh(; tess_path::String = Nothing, meshing::Dict = Dict(), 
custom_mesh_name::String = Nothing, force::Bool = false)

Creates and saves a mesh for the the provided tesselation file according to the meshing parameters. Returns the path to the mesh file.

If the force flag is set to true, overwriting of meshes is enabled. A custom mesh name can be given in the custom_mesh_name argument.

Arguments

tess_path::String=Nothing: Path to the tesselation file.

meshing::Dict=Dict(): Dictionary containg the meshing parameters.

custom_mesh_name::String=Nothing: Custom given mesh name.

force::Bool=false: Enables and disables overwriting of meshes.

NeperStructureGenerator.tesselate — Function

tesselate(; base_name::String = "crystal", parent_folder::String = pwd(), 
            tesselation::Dict = Dict())

Creates a tesselation with the respective orientation file according to the tesselation parameters. Returns the path to the created tesselation file.

Arguments

base_name::String="crystal": Name of the tesselation and orientation file.

parent_folder::String=pwd(): Root directory.

tesselation::Dict=Dict(): Dictionary containg the tesselation parameters.

NeperStructureGenerator.visualize_directory — Function

visualize_directory(output_dir::String, visualization_dir::String)

Creates visualizations for the tesselation and all meshes within the visualization_dir directory and saves it to the output_dir directory.

NeperStructureGenerator.visualize_mesh — Function

visualize_mesh(output_dir::String, mesh_path::String; mesh_name::String = Nothing)

Creates a visualization for the specified mesh and saves it to the output_dir directory.

Provide a path to a mesh file or a tesselation directory and a mesh name in order to visualize a mesh.

Arguments

output_dir::String: Path to directory for saving visualizations.

mesh_path::String: Path to the mesh file or directory containing the mesh.

mesh_name::String: Name of the mesh to visualize, if mesh_path is a directory.

all::Bool=false: Non public argument if all meshes should be visualized.

NeperStructureGenerator.visualize_tesselation — Function

visualize_tesselation(output_dir::String, tess_path::String)

Creates a visualizations for the specified tesselation and saves it to the output_dir directory.

Arguments

output_dir::String: Path to directory for saving visualizations.

tess_path::String: Path to the tesselation file or directory containing the tesselation.