Internals

NeperStructureGenerator.check_dicts_equal — Function

check_dicts_equal(dict::Dict, new_dict::Dict)

Returns true if the values of the keys in new_dict match the values of the keys in dict

NeperStructureGenerator.check_mesh_args — Function

check_mesh_args(file_path::String, new_mesh_dict::Dict, force::Bool = false)

Check if there already is a mesh with the same name and or same arguments.

If the force flag is set to true, overwriting of meshes is enabled.

Arguments

file_path::String: Path to the mesh file.

new_mesh_dictdict::Dict: Contains the parameters of the new mesh.

force::Bool=false: Enables and disables overwriting of meshes.

NeperStructureGenerator.create_cmdargs — Function

create_cmdargs(dict::Dict)

Create a string with the format "-<key1> <value1> -<key2> <value2> ...". The keys are not sorted.

NeperStructureGenerator.create_toml_data — Function

create_toml_data(name::String, dict::Dict, file_path::String; custom::Bool = false)

Create or update a TOML file with mesh data.

If the custom flag is set to true, the name parameter is used as the mesh name. Otherwise, a mesh name based on the number of existing entries is generated. The function saves the tesselation information in the "TESSELATION" section and the mesh information in the "MESHES" section of the TOML file. The TOML file is located in the same directory as the provided file_path.

Arguments

name::String: Provides the name of the tesselation dictionary in the toml file for tesselation and the path to the mesh file to infer the name for meshing.

dict::Dict: Contains information about the tesselation or meshing parameters.

file_path::String: Path to the tesselation or mesh file.

custom::Bool=false: Is set to true if a custom mesh name was given.

NeperStructureGenerator.generate_directory_name — Function

generate_directory_name(parent_folder::String, tesselation_settings::Dict)

Args: parentfolder (String): Root directory tesselationsettings (Dict): Dictionary containing the tesselation parameters

NeperStructureGenerator.name_from_dict — Function

name_from_dict(d::Dict)

Generate a string with the format "<key1><value1>_<key2><value2>_..." where the keys are sorted.

NeperStructureGenerator.generate_mesh_name — Function

Generates a mesh name based on the number of existing meshes.

Arguments

parent_folder::String: The path to the parent folder containing the input.toml file.

Returns

String: A generated mesh name in the format "Mesh_<number>", where <number> is the count of existing meshes plus one.

If a custom mesh name is not provided, this function counts the number of existing meshes and creates a new mesh name accordingly.